CSRIT_Project/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
cssselect2==0.7.0
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
lxml==4.9.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.8
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.16.4
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rest-framework-simplejwt==0.0.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.8.0
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.6
webencodings==0.5.1
whitenoise==6.6.0
wrapt==1.15.0
xhtml2pdf==0.2.11
